added disable enable teams. read and update user work

+++ 8RiverRodeo/src/main/java/com/skilldistillery/riverrodeo/services/UserServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import com.skilldistillery.riverrodeo.entities.Team;
 import com.skilldistillery.riverrodeo.entities.User;
+import com.skilldistillery.riverrodeo.repositories.TeamRepository;
 import com.skilldistillery.riverrodeo.repositories.UserRepository;
 
 @Service
@@ -15,6 +16,8 @@ public class UserServiceImpl implements UserService {
 	
 	@Autowired
 	private UserRepository userRepo;
+	@Autowired
+	private TeamRepository teamRepo;
 
 	@Override
 	public List<User> listAllUsers() {
@@ -32,15 +35,33 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public User updateUser(Integer userId, User user) {		
+	public User updateUser(Integer userId, User user, String teamname) {
+		System.out.println(user);
+		Team dbTeam = null;
+		if (user.getTeam() != null) {
+			System.out.println(user.getTeam().getId());
+			Optional<Team> optTeam = teamRepo.findById(user.getTeam().getId());
+			if (optTeam.isPresent()) {
+				dbTeam = optTeam.get();
+			}
+		}
+		Team loggedInTeam = teamRepo.findByName(teamname);
 		Optional<User> optUser = userRepo.findById(userId);
-		if (optUser.isPresent()) {
+		if (optUser.isPresent() && loggedInTeam != null) {
 			User managedUser = optUser.get();
 			if (managedUser != null) {
 				managedUser.setFirstName(user.getFirstName());
-				managedUser.setFishes(user.getFishes());
 				managedUser.setLastName(user.getLastName());
-				managedUser.setTeam(user.getTeam());
+				if (user.getFishes() != null) {
+					managedUser.setFishes(user.getFishes());					
+				}
+				if (dbTeam != null) {
+					// this grabs the team based on the id passed in on json
+					managedUser.setTeam(dbTeam);	
+				}else {
+					// this grabs the team from the user in the database
+					managedUser.setTeam(managedUser.getTeam());
+				}
 				return userRepo.saveAndFlush(managedUser);
 			}
 		}
@@ -48,9 +69,10 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public Boolean deleteUser(Integer userId) {
+	public Boolean deleteUser(Integer userId, String teamname) {
+		Team loggedInTeam = teamRepo.findByName(teamname);
 		Optional<User> optUser = userRepo.findById(userId);
-        if (optUser.isPresent()) {
+        if (optUser.isPresent() && loggedInTeam != null) {
 			User managedUser = optUser.get();
 			if (managedUser != null) {
 				userRepo.deleteById(userId);
